packages/llm/requirements.txt
faiss-cpu==1.9.0
fastapi==0.117.1
sentence-transformers==5.1.0
uvicorn==0.36.0
